+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fixed a wrong JNI method declaration which might cause "method not found" crash on some devices.
 * Fixed a bug that `Error` in the background async thread is not forwared to the caller thread.
+* Fixed a crash when an empty `Collection` is passed to `insert()`/`insertOrUpdate()` (#3103).
 
 ## 1.1.0