+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 * Fixed a bug that `Error` in the background async thread is not forwared to the caller thread.
 * Fixed a crash when an empty `Collection` is passed to `insert()`/`insertOrUpdate()` (#3103).
 
+### Internal
+
+* Updated Realm Core to 1.3.0.
+
 ## 1.1.0
 
 ### Bug fixes